Builds Scaden Model from scRNA data

The model is saved in a defined directory.

build_model_scaden(
  single_cell_object,
  celltype_labels,
  bulk_gene_expression = NULL,
  model_path = NULL,
  batch_size = 128,
  learning_rate = 1e-04,
  steps = 5000,
  var_cutoff = NULL,
  cells = 100,
  samples = 1000,
  dataset_name = "scaden",
  verbose = FALSE
)

Arguments

single_cell_object
A matrix or dataframe with the single-cell data. Rows are genes, columns are samples.

celltype_labels
Vector of celltype labels. Order corresponds to rows in single_cell_object matrix.

bulk_gene_expression
Matrix of bulk RNA expression. (rows :genes, columns: samples)

model_path
Path where model directory should be created. (optional)

batch_size
Training of model: Batch size to use for training. (default: 128)

learning_rate
Training of model: Learning rate used for training. (default: 0.0001)

steps
Training of model: Number of training steps (default: 5000)

var_cutoff
Training data processing: Filter out genes with a variance less than the specified cutoff. A low cutoff is recommended,this should only remove genes that are obviously uninformative. (default NULL)

cells
Bulk simulation: Number of cells per sample (default: 100)

samples
Bulk simulation: Number of samples to simulate (default: 1000)

dataset_name
Bulk simulation: Name of simulated dataset (default scaden)

verbose
If true: Command line output of Scaden is shown (default: false)

Value

Scaden model